+++ docs/include/samplecode/SoundTouchClient/GetBlueToothInfo.py
@@ -1,0 +1,28 @@
+from bosesoundtouchapi import *
+from bosesoundtouchapi.models import *
+from bosesoundtouchapi.uri import *
+
+try:
+    
+    # create SoundTouch device instance.
+    device:SoundTouchDevice = SoundTouchDevice("192.168.1.131")
+            
+    # create SoundTouch client instance from device.
+    client:SoundTouchClient = SoundTouchClient(device)
+
+    # get real-time configuration from the device.
+    config:BlueToothInfo = client.GetBlueToothInfo()
+    print(config.ToString())
+
+    # get cached configuration, refreshing from device if needed.
+    config:BlueToothInfo = client.GetBlueToothInfo(False)
+    print("\nCached configuration:\n%s" % config.ToString())
+
+    # get cached configuration directly from the configuration manager dictionary.
+    if SoundTouchNodes.bluetoothInfo.Path in client.ConfigurationCache:
+        config:BlueToothInfo = client.ConfigurationCache[SoundTouchNodes.bluetoothInfo.Path]
+        print("\nCached configuration, direct:\n%s" % config.ToString())
+        
+except Exception as ex:
+
+    print("** Exception: %s" % str(ex))